Remove old NewGuestUser notification
